Updated polymorph init and recast timers.
https://bitbucket.org/looking4group_b2tbc/looking4group/issues/2549/greater-polymorph-kiggler-the-crazed

The proposed recast timer from the report seems reasonable, although it is possible that it should be between 10 and 20 seconds rather than 10 and 25.

As for the init timer.. It should most likely be longer than the recast timer, but due to a lack of evidence I've set it to the same as the recast timer. Instead of the old value of 5 seconds, which I can not find support of anywhere.

The report also mentions that polymorphs are supposed to only be cast at the primary target, I have only touched the timers.