Updated timer for Tito's summon
https://bitbucket.org/looking4group_b2tbc/looking4group/issues/2603/dorothy-does-not-summon-tito

Last reply provides sources indicating that the timer is supposed to be a random interval that includes 37 and 40. This commit sets the interval to 36-41, as it is hard to pinpoint the exact time the spell is cast. Only having 2 samples is also not optimal, the interval should possibly be larger (in both directions, or just one).