Update boss_tailonking_ikiss.cpp
https://github.com/cmangos/mangos-tbc/blob/master/src/scriptdev2/scripts/outland/auchindoun/sethekk_halls/boss_talon_king_ikiss.cpp

https://github.com/NostalriusTBC/Core/blob/master/src/scriptdev2/scripts/outland/auchindoun/sethekk_halls/boss_talon_king_ikiss.cpp

https://github.com/OregonCore/OregonCore/blob/master/src/scripts/Outland/Auchindoun/sethekk_halls/boss_tailonking_ikiss.cpp

http://wowwiki.wikia.com/wiki/Talon_King_Ikiss?direction=next&oldid=514381

As of Feb. 23 2007, he now has a fairly high chance to chain blink and cast his explosion repeatedly. (5sec casttime) This means that if your party members are hiding behind a pillar when he casts arcane explosion, he may blink behind that pillar and begin casting it again. A GM has confirmed that this is an intended change.

This was later nerfed to flat % HP Values.