Update boss_captain_skarloc.cpp
http://www.wowhead.com/npc=17862/captain-skarloc#comments
http://www.wowhead.com/npc=17862/captain-skarloc#abilities:mode=heroic
http://www.wowhead.com/spell=38385/consecration

On Heroic, the boss himself has only one serious threat, and that's his Consecration. Not only is it nasty, but he can cast two of them at once. If several members of your party are in it, your healer can't keep them all up.

Simply put, the tank must keep moving around to stay out of consecration. If the tank stays still, everyone gets hit by double consecration, and everyone dies. Melee attackers must not wait for the tank, but move out of it on their own.